allah jne ki korsi vai

+++ app/Http/Controllers/Admin/AdminProfileController.php
@@ -9,40 +9,34 @@ use Illuminate\Support\Facades\Storage;
 
 class AdminProfileController extends Controller
 {
-    function viewProfile()
-    {
-        return view('backend.profile.index');
-    }
+    //
+}
 
-    function updateProfile(Request $request)
-    {
-        // Validation
-        $request->validate([
-            'profile_img' => 'nullable|mimes:jpg,jpeg,png,webp|max:5048', // kb,
-            'name' => 'required|min:2',
-        ]);
+ function updateProfile(Request $request)
+{
+    $request->validate([
+        'profile_img' => 'nullable|mimes:jpg,jpeg,png,webp|max:5048',
+        'name' => 'required|min:2',
+    ]);
 
+    $fileName = auth()->user()->profile_img;
 
-        // User Update
-        $fileName = null;
-        if ($request->hasFile('profile_img')) {
-            // Previous Image exists
-            // If exists Prev Delete
-            if(Storage::disk('public')->exists(auth()->user()->profile_img)){
-                Storage::disk('public')->delete(auth()->user()->profile_img);
-            }
+    if ($request->hasFile('profile_img')) {
 
-
-            $fileName  = $request->profile_img->store('profile', 'public');
+        if (
+            auth()->user()->profile_img &&
+            Storage::disk('public')->exists(auth()->user()->profile_img)
+        ) {
+            Storage::disk('public')->delete(auth()->user()->profile_img);
         }
 
-
-        // User DB Update
-
-        $authUser = User::find(auth()->user()->id)->update([
-            'name' => $request->name,
-            'profile_img' => $fileName,
-        ]);
-        return back();
+        $fileName = $request->file('profile_img')->store('profile', 'public');
     }
+
+    User::find(auth()->id())->update([
+        'name' => $request->name,
+        'profile_img' => $fileName,
+    ]);
+
+    return back()->with('success', 'Profile updated successfully.');
 }
